Move transaction translate functions to translate-cart file
They are not types and belong outside the types file.
