+++ client/my-sites/checkout/composite-checkout/lib/translate-cart.ts
@@ -3,6 +3,7 @@
  */
 import { translate } from 'i18n-calypso';
 import { ResponseCart, ResponseCartProduct } from '@automattic/shopping-cart';
+import debugFactory from 'debug';
 
 /**
  * Internal dependencies
@@ -22,6 +23,16 @@ import { isPlan, isDomainTransferProduct, isDomainProduct } from 'calypso/lib/pr
 import { isRenewal } from 'calypso/lib/cart-values/cart-items';
 import doesValueExist from './does-value-exist';
 import doesPurchaseHaveFullCredits from './does-purchase-have-full-credits';
+import type { DomainContactDetails } from '../types/backend/domain-contact-details-components';
+import type {
+	WPCOMTransactionEndpointCart,
+	WPCOMTransactionEndpointCartItem,
+	WPCOMTransactionEndpointRequestPayload,
+	TransactionRequestWithLineItems,
+} from '../types/transaction-endpoint';
+import { isGSuiteProductSlug } from 'calypso/lib/gsuite';
+
+const debug = debugFactory( 'calypso:composite-checkout:translate-cart' );
 
 /**
  * Translate a cart object as returned by the WPCOM cart endpoint to
@@ -265,4 +276,155 @@ function translateReponseCartProductToWPCOMCartItem(
 
 export function getNonProductWPCOMCartItemTypes(): string[] {
 	return [ 'tax', 'coupon', 'total', 'subtotal', 'credits', 'savings' ];
+}
+
+// Create cart object as required by the WPCOM transactions endpoint
+// '/me/transactions/': WPCOM_JSON_API_Transactions_Endpoint
+export function createTransactionEndpointCartFromLineItems( {
+	siteId,
+	couponId,
+	country,
+	postalCode,
+	subdivisionCode,
+	items,
+	contactDetails,
+}: {
+	siteId: string;
+	couponId?: string;
+	country: string;
+	postalCode: string;
+	subdivisionCode?: string;
+	items: WPCOMCartItem[];
+	contactDetails: DomainContactDetails | null;
+} ): WPCOMTransactionEndpointCart {
+	debug( 'creating cart from items', items );
+
+	const currency: string = items.reduce(
+		( firstValue: string, item ) => firstValue || item.amount.currency,
+		''
+	);
+
+	return {
+		blog_id: siteId || '0',
+		cart_key: siteId || 'no-site',
+		create_new_blog: siteId ? false : true,
+		coupon: couponId || '',
+		currency: currency || '',
+		temporary: false,
+		extra: [],
+		products: items
+			.filter( ( product ) => ! getNonProductWPCOMCartItemTypes().includes( product.type ) )
+			.map( ( item ) => addRegistrationDataToGSuiteItem( item, contactDetails ) )
+			.map( createTransactionEndpointCartItemFromLineItem ),
+		tax: {
+			location: {
+				country_code: country,
+				postal_code: postalCode,
+				subdivision_code: subdivisionCode,
+			},
+		},
+	};
+}
+
+function createTransactionEndpointCartItemFromLineItem(
+	item: WPCOMCartItem
+): WPCOMTransactionEndpointCartItem {
+	return {
+		product_id: item.wpcom_meta?.product_id,
+		meta: item.wpcom_meta?.meta,
+		currency: item.amount.currency,
+		volume: item.wpcom_meta?.volume ?? 1,
+		quantity: item.wpcom_meta?.quantity ?? null,
+		extra: item.wpcom_meta?.extra,
+	} as WPCOMTransactionEndpointCartItem;
+}
+
+function addRegistrationDataToGSuiteItem(
+	item: WPCOMCartItem,
+	contactDetails: DomainContactDetails | null
+): WPCOMCartItem {
+	if ( ! isGSuiteProductSlug( item.wpcom_meta?.product_slug ) ) {
+		return item;
+	}
+	return {
+		...item,
+		wpcom_meta: {
+			...item.wpcom_meta,
+			extra: { ...item.wpcom_meta.extra, google_apps_registration_data: contactDetails },
+		},
+	} as WPCOMCartItem;
+}
+
+export function createTransactionEndpointRequestPayloadFromLineItems( {
+	siteId,
+	couponId,
+	country,
+	state,
+	postalCode,
+	subdivisionCode,
+	city,
+	address,
+	streetNumber,
+	phoneNumber,
+	document,
+	deviceId,
+	domainDetails,
+	items,
+	paymentMethodType,
+	paymentMethodToken,
+	paymentPartnerProcessorId,
+	storedDetailsId,
+	name,
+	email,
+	cancelUrl,
+	successUrl,
+	idealBank,
+	tefBank,
+	pan,
+	gstin,
+	nik,
+}: TransactionRequestWithLineItems ): WPCOMTransactionEndpointRequestPayload {
+	return {
+		cart: createTransactionEndpointCartFromLineItems( {
+			siteId,
+			couponId: couponId || getCouponIdFromProducts( items ),
+			country,
+			postalCode,
+			subdivisionCode,
+			items: items.filter( ( item ) => item.type !== 'tax' ),
+			contactDetails: domainDetails || {},
+		} ),
+		domainDetails,
+		payment: {
+			paymentMethod: paymentMethodType,
+			paymentKey: paymentMethodToken,
+			paymentPartner: paymentPartnerProcessorId,
+			storedDetailsId,
+			name,
+			email,
+			country,
+			countryCode: country,
+			state,
+			postalCode,
+			zip: postalCode, // TODO: do we need this in addition to postalCode?
+			city,
+			address,
+			streetNumber,
+			phoneNumber,
+			document,
+			deviceId,
+			successUrl,
+			cancelUrl,
+			idealBank,
+			tefBank,
+			pan,
+			gstin,
+			nik,
+		},
+	};
+}
+
+function getCouponIdFromProducts( items: WPCOMCartItem[] ): string | undefined {
+	const couponItem = items.find( ( item ) => item.type === 'coupon' );
+	return couponItem?.wpcom_meta?.couponCode;
 }
